requirements.txt
requests
beautifulsoup4
lxml
pillow
pystray
plyer
